# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "智能问数据分析系统 API",
    "version": "0.1.0",
    "description": "自然语言问数 / 分析报告 / 数据源与知识管理后端接口。除标注「公开」的端点外均需 Bearer JWT；写操作与管理员端点在描述中注明角色要求（RBAC：ADMIN / ANALYST / VIEWER）。错误响应统一为 { error: string, code?: string }，code 取值见 server/errorCodes.ts。"
  },
  "servers": [
    {
      "url": "/",
      "description": "同源部署（开发默认 http://127.0.0.1:3000）"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/system/models": {
      "get": {
        "tags": [
          "System"
        ],
        "operationId": "listModels",
        "summary": "可选模型目录（Ollama 实时列已装模型，qwen/gemini 按密钥列入）",
        "responses": {
          "200": {
            "description": "模型列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "models": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        }
      }
    },
    "/api/system/llm-usage": {
      "get": {
        "tags": [
          "System"
        ],
        "operationId": "getLlmUsage",
        "summary": "LLM 用量统计（近 N 天按引擎/模型聚合；仅 ADMIN）",
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 7,
              "minimum": 1,
              "maximum": 90
            },
            "description": "统计天数"
          }
        ],
        "responses": {
          "200": {
            "description": "用量聚合"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
    "/api/datasources": {
      "get": {
        "tags": [
          "DataSources"
        ],
        "operationId": "listDatasources",
        "summary": "数据源列表（ADMIN 全量；ANALYST/VIEWER 仅授权范围）",
        "responses": {
          "200": {
            "description": "数据源列表"
          }
        }
      },
      "post": {
        "tags": [
          "DataSources"
        ],
        "operationId": "createDatasource",
        "summary": "新建数据源（仅 ADMIN；凭据加密存储）",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DatasourceInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "创建成功（含连通性校验结果）"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/api/datasources/import-file": {
      "post": {
        "tags": [
          "DataSources"
        ],
        "operationId": "importFileDatasource",
        "summary": "上传 CSV/Excel/JSON 文件导入为数据源（仅 ADMIN）：服务端解析真实数据落应用库物理表 upl_*，问数/报表/指标/灵活查询走真实执行；≤7MB、≤2万行、≤100列（超限截断），敏感列自动剔除",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "fileType",
                  "fileName",
                  "fileBase64"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "数据源显示名（可选，默认取文件名）"
                  },
                  "fileType": {
                    "type": "string",
                    "enum": [
                      "csv",
                      "xlsx",
                      "json"
                    ]
                  },
                  "fileName": {
                    "type": "string",
                    "description": "原始文件名"
                  },
                  "fileBase64": {
                    "type": "string",
                    "description": "文件内容 base64（文件本体 ≤7MB）"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "导入成功：返回 dataSource 与 stats（rows/columns/droppedSensitive/truncated）"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
    "/api/conversations": {
      "get": {
        "tags": [
          "Conversations"
        ],
        "operationId": "listConversations",
        "summary": "当前用户对话历史（按数据源隔离，支持关键词检索；ADMIN/ANALYST）",
        "parameters": [
          {
            "name": "dataSourceId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "问题/结论关键词"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "conversations 列表"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
    "/api/report/export": {
      "post": {
        "tags": [
          "Reports"
        ],
        "operationId": "exportReport",
        "summary": "导出报告（PDF/PPTX/Markdown；body 上限 20MB 含图表 base64）",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "format"
                ],
                "properties": {
                  "format": {
                    "type": "string",
                    "enum": [
                      "pdf",
                      "pptx",
                      "markdown"
                    ]
                  },
                  "title": {
                    "type": "string"
                  },
                  "sections": {
                    "type": "array",
                    "items": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "文件流"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
    "/api/knowledge": {
      "get": {
        "tags": [
          "Knowledge"
        ],
        "operationId": "listKnowledge",
        "summary": "知识文档列表（支持关键词检索）",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dataSourceId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "文档列表"
          }
        }
      },
      "post": {
        "tags": [
          "Knowledge"
        ],
        "operationId": "createKnowledge",
        "summary": "新建知识文档（仅 ADMIN；供问数语义增强）",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "title",
                  "content"
                ],
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "content": {
                    "type": "string"
                  },
                  "dataSourceId": {
                    "type": "string"
                  },
                  "tags": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "创建成功"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
    "/api/sql-examples": {
      "get": {
        "tags": [
          "SqlExamples"
        ],
        "operationId": "listSqlExamples",
        "summary": "SQL 示例列表（问数 few-shot 语料；支持 dataSourceId/q 过滤）",
        "parameters": [
          {
            "name": "dataSourceId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "示例列表"
          }
        }
      },
      "post": {
        "tags": [
          "SqlExamples"
        ],
        "operationId": "createSqlExample",
        "summary": "新建 SQL 示例（仅 ADMIN）",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "question",
                  "sql"
                ],
                "properties": {
                  "question": {
                    "type": "string"
                  },
                  "sql": {
                    "type": "string"
                  },
                  "dataSourceId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "创建成功"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
    "/api/sql-examples/export": {
      "get": {
        "tags": [
          "SqlExamples"
        ],
        "operationId": "exportSqlExamples",
        "summary": "SQL 样例库导出（仅 ADMIN；JSON 备份，含问题/SQL/来源标记）",
        "parameters": [
          {
            "name": "dataSourceId",
            "in": "query",
            "required": true,
            "description": "数据源 ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          },
          "Forbidden": {
            "$ref": "#/components/responses/Forbidden"
          },
          "NotFound": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/api/skills": {
      "get": {
        "tags": [
          "Skills"
        ],
        "operationId": "listSkills",
        "summary": "技能库列表（含公共技能与我创建的技能）",
        "responses": {
          "200": {
            "description": "技能列表"
          }
        }
      },
      "post": {
        "tags": [
          "Skills"
        ],
        "operationId": "createSkill",
        "summary": "新建技能（提示词模板 + 参数）",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "创建成功"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
...
    "/api/access-requests": {
      "post": {
        "tags": [
          "AccessRequests"
        ],
        "operationId": "createAccessRequest",
        "summary": "提交数据源访问申请",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "dataSourceId"
                ],
                "properties": {
                  "dataSourceId": {
                    "type": "string"
                  },
                  "reason": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功"
          },
          "BadRequest": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      },
      "get": {
        "tags": [
          "AccessRequests"
        ],
        "operationId": "listAccessRequests",
        "summary": "全部访问申请（仅 ADMIN）",
        "responses": {
          "200": {
            "description": "成功"
          },
          "Forbidden": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
    "/api/knowledge-external": {
      "get": {
        "tags": [
          "ExternalKnowledge"
        ],
        "operationId": "listExternalKb",
        "summary": "外部知识库配置列表（仅 ADMIN；api_key 不出明文）",
        "responses": {
          "200": {
            "description": "成功"
          },
          "Forbidden": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      },
      "post": {
        "tags": [
          "ExternalKnowledge"
        ],
        "operationId": "createExternalKb",
        "summary": "新增外部知识库配置（仅 ADMIN）",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "endpoint"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "endpoint": {
                    "type": "string"
                  },
                  "apiKey": {
                    "type": "string"
                  },
                  "enabled": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功"
          },
          "BadRequest": {
            "$ref": "#/components/responses/BadRequest"
          },
          "Forbidden": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
    "/api/knowledge/export": {
      "get": {
        "tags": [
          "Knowledge"
        ],
        "operationId": "exportKnowledge",
        "summary": "知识库导出（JSON 备份，含元数据与版本信息）",
        "responses": {
          "200": {
            "description": "成功"
          },
          "Forbidden": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
    "/api/knowledge/import": {
      "post": {
        "tags": [
          "Knowledge"
        ],
        "operationId": "importKnowledge",
        "summary": "知识库导入（合并策略 replace/append/skip；body 上限 10mb）",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "data"
                ],
                "properties": {
                  "data": {
                    "type": "object"
                  },
                  "mergeStrategy": {
                    "type": "string",
                    "enum": [
                      "replace",
                      "append",
                      "skip"
                    ]
                  },
                  "dryRun": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功"
          },
          "BadRequest": {
            "$ref": "#/components/responses/BadRequest"
          },
          "Forbidden": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
    "/api/report-templates": {
      "get": {
        "tags": [
          "Reports"
        ],
        "operationId": "listReportTemplates",
        "summary": "报告模板列表（含预置模板）",
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      },
      "post": {
        "tags": [
          "Reports"
        ],
        "operationId": "createReportTemplate",
        "summary": "新建报告模板（仅 ADMIN）",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "templateContent"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "templateContent": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功"
          },
          "BadRequest": {
            "$ref": "#/components/responses/BadRequest"
          },
          "Forbidden": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
    "/api/saved-reports": {
      "get": {
        "tags": [
          "Reports"
        ],
        "operationId": "listSavedReports",
        "summary": "历史决策报表列表（所有登录用户可见，可按数据源过滤）",
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      },
      "post": {
        "tags": [
          "Reports"
        ],
        "operationId": "createSavedReport",
        "summary": "保存决策报表（ADMIN/ANALYST；report_id 冲突 409 幂等）",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "report"
                ],
                "properties": {
                  "report": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功"
          },
          "BadRequest": {
            "$ref": "#/components/responses/BadRequest"
          },
          "Forbidden": {
            "$ref": "#/components/responses/Forbidden"
          },
          "Conflict": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      }
    },
...
    "/api/dashboard-widgets": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "operationId": "listDashboardWidgets",
        "summary": "看板固化图表列表（所有登录用户可见，按 sort_order 升序）",
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      },
      "post": {
        "tags": [
          "Dashboard"
        ],
        "operationId": "pinDashboardWidget",
        "summary": "固化图表至看板（ADMIN/ANALYST；widget_id 冲突 409 幂等）",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "widget"
                ],
                "properties": {
                  "widget": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功"
          },
          "BadRequest": {
            "$ref": "#/components/responses/BadRequest"
          },
          "Forbidden": {
            "$ref": "#/components/responses/Forbidden"
          },
          "Conflict": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
    "responses": {
      "BadRequest": {
        "description": "参数错误（INVALID_INPUT）",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "Forbidden": {
        "description": "角色不足或被强制改密拦截（FORBIDDEN / PASSWORD_CHANGE_REQUIRED）",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "资源不存在或无权访问（NOT_FOUND，越权访问同样返回 404 防枚举）",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "RateLimited": {
        "description": "触发限流（RATE_LIMITED）",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Internal": {
        "description": "服务端错误（INTERNAL_ERROR）",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Conflict": {
        "description": "资源冲突（如唯一标识已存在；幂等场景可安全忽略）",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "schemas": {
...
      "DatasourceInput": {
        "type": "object",
        "required": [
          "name",
          "type",
          "config"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "mysql",
              "postgres",
              "csv"
            ]
          },
          "config": {
            "type": "object",
            "description": "连接配置（host/port/database/username/password 或 CSV 上传参数）"
          },
          "description": {
            "type": "string"
          }
        }
      }
    }
  }
}